move AppProvieder to a different file
