+++ src/lib/types/types.d.ts
@@ -13,6 +13,7 @@ export type PageDataType = {
 export interface AppContextType {
     theme: string;
     setTheme: (theme: string) => void;
-    pageData: PageDataType;
+    pageData: PageDataType | null;
+    
     setPageData: React.Dispatch<React.SetStateAction<PageDataType | null>>;
 };
