+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import App from './pages/App/';
-import { AppProvider } from './context/AppContext';
+import { AppProvider } from './context/AppProvider';
 import './styles/index.css';
 
 createRoot(document.getElementById('root')!).render(
